Config
Config機能を追加
何でもかんでも弄れるぞ！
