+++ AllergyProfileAPI/src/main/java/com/github/distriful5061/AllergyProfile/mixins/ConfigJson.java
@@ -1,0 +1,9 @@
+package com.github.distriful5061.AllergyProfile.mixins;
+
+public class ConfigJson {
+    public int port;
+
+    public ConfigJson(int port) {
+        this.port = port;
+    }
+}
