+++ AllergyProfileAPI/src/main/resources/config.json
@@ -1,0 +1,3 @@
+{
+  "port": 8080
+}
