+++ README.md
@@ -2,4 +2,4 @@
 
 An application that allows users to easily check allergy information using URLs or QR codes.
 
-From (idee)[https://www.ideee.tech/ideas/2025]
+From https://www.ideee.tech/ideas/2025
